Don't store en passant file
Previously, a board object's bitfield stored en passant rights as a file (0 - 7) or "no ep file" (8).

Now, the bitfield stores a single bit to indicate if en passant rights are present for the player to move, and if so, the en passant file is read from the previous move info (get_end_file()).